Re-organise fink-utils
